+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ## Unreleased
 
+### Added
+
+- `epts_code` validator
+
 ### Changed
 
 - Rules of `vin_code`, `body_code`, `chassis_code`, `grz_code`, `sts_code` and `pts_code` validators